Switch from qemu to vmrt
* use vmrt instead of qemu
* build.rs for init-container
* cargo deb metadata
* bump version to 0.2.0
