--- init-container/.gitignore
@@ -1,5 +1,0 @@
-unpacked_kernel
-vmlinuz-virt
-init
-initramfs.cpio.gz
-squashfs_drive
