--- init-container/tests/.gitignore
@@ -1,1 +1,0 @@
-cyclic_buffer
