--- poc/runtime/bios-256k.bin
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:8fdc1ec7bdedcd113572a9a983509a12f344bd98b7e74ed1988e7e13e259f44a
-size 262144
