--- poc/runtime/initramfs-virt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:272441389a20e761f8d5fa7d69a714289b469f27e8ebddeaa07bfb7bcaf719b6
-size 5386251
