--- poc/runtime/kvmvapic.bin
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:cdf057a71b07e3b52b19cbe210bdefa59250d01a9810b960f7fe1f98eed95a27
-size 9216
