--- poc/runtime/linuxboot_dma.bin
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9c49e255340c78fc12e54ed043462bca02fb7fca29b7cfab62ff88a5344b6950
-size 1536
