--- poc/runtime/vmlinuz-virt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6f58942db75ed5c918517c89a6f15dbc69c70b7a402ad26681ecfde2f843e67e
-size 4526240
