--- poc/runtime/vmrt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c9288c0fcce6a15447f3fe2bdcf04e36306f6441eeb834ef572e65bbdcdefe86
-size 8395016
