+++ runtime/init-container/tests/.gitignore
@@ -1,0 +1,1 @@
+cyclic_buffer
